+++ packages/affine/shared/src/services/theme-service.ts
@@ -1,0 +1,213 @@
+import { type Color, ColorScheme } from '@blocksuite/affine-model';
+import {
+  type BlockStdScope,
+  Extension,
+  type ExtensionType,
+  StdIdentifier,
+} from '@blocksuite/block-std';
+import { type Container, createIdentifier } from '@blocksuite/global/di';
+import { signal, type Signal } from '@preact/signals-core';
+import {
+  type AffineCssVariables,
+  combinedDarkCssVariables,
+  combinedLightCssVariables,
+} from '@toeverything/theme';
+import { unsafeCSS } from 'lit';
+
+import { DocModeProvider, type DocModeService } from './doc-mode-service.js';
+
+const TRANSPARENT = 'transparent';
+
+export const ThemeExtensionIdentifier = createIdentifier<ThemeExtension>(
+  'AffineThemeExtension'
+);
+
+export interface ThemeExtension {
+  getAppTheme?: () => Signal<ColorScheme>;
+  getEdgelessTheme?: (docId?: string) => Signal<ColorScheme>;
+}
+
+export function OverrideThemeExtension(service: ThemeExtension): ExtensionType {
+  return {
+    setup: di => {
+      di.override(ThemeExtensionIdentifier, () => service);
+    },
+  };
+}
+
+export const ThemeProvider = createIdentifier<ThemeService>(
+  'AffineThemeProvider'
+);
+
+export class ThemeService extends Extension {
+  app$: Signal<ColorScheme>;
+
+  edgeless$: Signal<ColorScheme>;
+
+  get theme() {
+    return this.docMode.getEditorMode() === 'page'
+      ? this.app$.peek()
+      : this.edgeless$.peek();
+  }
+
+  get theme$() {
+    return this.docMode.getEditorMode() === 'page' ? this.app$ : this.edgeless$;
+  }
+
+  constructor(
+    private std: BlockStdScope,
+    private docMode: DocModeService
+  ) {
+    super();
+    const extension = this.std.getOptional(ThemeExtensionIdentifier);
+    this.app$ = extension?.getAppTheme?.() || themeObserver.theme$;
+    this.edgeless$ = extension?.getEdgelessTheme?.() || themeObserver.theme$;
+  }
+
+  static override setup(di: Container) {
+    di.addImpl(ThemeProvider, ThemeService, [StdIdentifier, DocModeProvider]);
+  }
+
+  /**
+   * Generates a CSS's color property with `var` or `light-dark` functions.
+   *
+   * Sometimes used to set the frame/note background.
+   *
+   * @param color - A color value.
+   * @param fallback  - If color value processing fails, it will be used as a fallback.
+   * @returns - A color property string.
+   *
+   * @example
+   *
+   * ```
+   * `rgba(255,0,0)`
+   * `#fff`
+   * `light-dark(#fff, #000)`
+   * `var(--affine-palette-shape-blue)`
+   * ```
+   */
+  generateColorProperty(
+    color: Color,
+    fallback = 'transparent',
+    theme = this.theme
+  ) {
+    let result: string | undefined = undefined;
+
+    if (typeof color === 'object') {
+      result = color[theme] ?? color.normal;
+    } else {
+      result = color;
+    }
+    if (!result) {
+      result = fallback;
+    }
+    if (result.startsWith('--')) {
+      return result.endsWith(TRANSPARENT) ? TRANSPARENT : `var(${result})`;
+    }
+
+    return result ?? TRANSPARENT;
+  }
+
+  /**
+   * Gets a color with the current theme.
+   *
+   * @param color - A color value.
+   * @param fallback - If color value processing fails, it will be used as a fallback.
+   * @param real - If true, it returns the computed style.
+   * @returns - A color property string.
+   *
+   * @example
+   *
+   * ```
+   * `rgba(255,0,0)`
+   * `#fff`
+   * `--affine-palette-shape-blue`
+   * ```
+   */
+  getColorValue(
+    color: Color,
+    fallback = TRANSPARENT,
+    real = false,
+    theme = this.theme
+  ) {
+    let result: string | undefined = undefined;
+
+    if (typeof color === 'object') {
+      result = color[theme] ?? color.normal;
+    } else {
+      result = color;
+    }
+    if (!result) {
+      result = fallback;
+    }
+    if (real && result.startsWith('--')) {
+      result = result.endsWith(TRANSPARENT)
+        ? TRANSPARENT
+        : this.getCssVariableColor(result);
+    }
+
+    return result ?? TRANSPARENT;
+  }
+
+  getCssVariableColor(property: string, theme = this.theme) {
+    if (property.startsWith('--')) {
+      if (property.endsWith(TRANSPARENT)) {
+        return TRANSPARENT;
+      }
+      const key = property as keyof AffineCssVariables;
+      const color =
+        theme === ColorScheme.Dark
+          ? combinedDarkCssVariables[key]
+          : combinedLightCssVariables[key];
+      return color;
+    }
+    return property;
+  }
+}
+
+class ThemeObserver {
+  private observer: MutationObserver;
+
+  theme$ = signal(ColorScheme.Light);
+
+  constructor() {
+    const COLOR_SCHEMES: string[] = Object.values(ColorScheme);
+    this.observer = new MutationObserver(() => {
+      const mode = document.documentElement.dataset.theme;
+      if (!mode) return;
+      if (!COLOR_SCHEMES.includes(mode)) return;
+      if (mode === this.theme$.value) return;
+
+      this.theme$.value = mode as ColorScheme;
+    });
+    this.observer.observe(document.documentElement, {
+      attributes: true,
+      attributeFilter: ['data-theme'],
+    });
+  }
+
+  destroy() {
+    this.observer.disconnect();
+  }
+}
+
+const themeObserver = new ThemeObserver();
+
+const toolbarColorKeys: Array<keyof AffineCssVariables> = [
+  '--affine-background-overlay-panel-color',
+  '--affine-background-primary-color',
+  '--affine-background-tertiary-color',
+  '--affine-icon-color',
+  '--affine-icon-secondary',
+  '--affine-border-color',
+  '--affine-divider-color',
+  '--affine-text-primary-color',
+];
+
+export const lightToolbarStyles = toolbarColorKeys.map(
+  key => `${key}: ${unsafeCSS(combinedLightCssVariables[key])};`
+);
+
+export const darkToolbarStyles = toolbarColorKeys.map(
+  key => `${key}: ${unsafeCSS(combinedDarkCssVariables[key])};`
+);