+++ packages/blocks/src/frame-block/frame-block.ts
@@ -2,7 +2,8 @@ import type { BlockStdScope } from '@blocksuite/block-std';
 import type { Doc } from '@blocksuite/store';
 
 import { ColorScheme, FrameBlockModel } from '@blocksuite/affine-model';
-import { ThemeObserver, unsafeCSSVarV2 } from '@blocksuite/affine-shared/theme';
+import { ThemeProvider } from '@blocksuite/affine-shared/services';
+import { unsafeCSSVarV2 } from '@blocksuite/affine-shared/theme';
 import {
   docContext,
   GfxBlockComponent,
@@ -14,7 +15,7 @@ import { GfxControllerIdentifier } from '@blocksuite/block-std/gfx';
 import { SignalWatcher, WithDisposable } from '@blocksuite/global/utils';
 import { Bound, type SerializedXYWH } from '@blocksuite/global/utils';
 import { consume } from '@lit/context';
-import { cssVarV2 } from '@toeverything/theme/v2';
+import { cssVarV2, themeToVar } from '@toeverything/theme/v2';
 import { css, html, nothing } from 'lit';
 import { query, state } from 'lit/decorators.js';
 import { styleMap } from 'lit/directives/style-map.js';
@@ -73,26 +74,22 @@ export class EdgelessFrameTitle extends SignalWatcher(
   private _cachedWidth = 0;
 
   get colors() {
-    let backgroundColor = ThemeObserver.getColorValue(
-      this.model.background,
-      undefined,
-      true
-    );
+    let backgroundColor = this.std
+      .get(ThemeProvider)
+      .getColorValue(this.model.background, undefined, true);
     if (isTransparent(backgroundColor)) {
-      backgroundColor = ThemeObserver.getPropertyValue(
-        cssVarV2('edgeless/frame/background/white').replace(
-          /var\((--.*)\)/,
-          '$1'
-        )
-      );
+      backgroundColor = this.std
+        .get(ThemeProvider)
+        .getCssVariableColor(themeToVar('edgeless/frame/background/white'));
     }
 
     const { r, g, b, a } = parseStringToRgba(backgroundColor);
 
+    const theme = this.std.get(ThemeProvider).theme;
     let textColor: string;
     {
       let rPrime, gPrime, bPrime;
-      if (ThemeObserver.instance.mode$.value === ColorScheme.Light) {
+      if (theme === ColorScheme.Light) {
         rPrime = 1 - a + a * r;
         gPrime = 1 - a + a * g;
         bPrime = 1 - a + a * b;
@@ -344,11 +341,10 @@ export class FrameBlockComponent extends GfxBlockComponent<FrameBlockModel> {
   }
 
   override renderGfxBlock() {
-    const { model, showBorder, rootService } = this;
-    const backgroundColor = ThemeObserver.generateColorProperty(
-      model.background,
-      '--affine-platte-transparent'
-    );
+    const { model, showBorder, rootService, std } = this;
+    const backgroundColor = std
+      .get(ThemeProvider)
+      .generateColorProperty(model.background, '--affine-platte-transparent');
     const _isNavigator =
       this.gfx.tool.currentToolName$.value === 'frameNavigator';
     const frameIndex = rootService.layer.getZIndex(model);