+++ packages/blocks/src/surface-block/mini-mindmap/spec.ts
@@ -4,6 +4,10 @@ import type { z } from 'zod';
 import { SurfaceBlockSchema } from '@blocksuite/affine-block-surface';
 import { RootBlockSchema } from '@blocksuite/affine-model';
 import {
+  DocModeService,
+  ThemeService,
+} from '@blocksuite/affine-shared/services';
+import {
   BlockViewExtension,
   type ExtensionType,
   FlavourExtension,
@@ -14,10 +18,11 @@ import { MindmapService } from './minmap-service.js';
 import { MindmapSurfaceBlockService } from './surface-service.js';
 
 export const MiniMindmapSpecs: ExtensionType[] = [
+  DocModeService,
+  ThemeService,
   FlavourExtension('affine:page'),
   MindmapService,
   BlockViewExtension('affine:page', literal`mini-mindmap-root-block`),
-
   FlavourExtension('affine:surface'),
   MindmapSurfaceBlockService,
   BlockViewExtension('affine:surface', literal`mini-mindmap-surface-block`),